ec2: sort and print all images by creation date
This fixes a bug with fedora-36 that always choose the Beta-xxx image.
This is the logical thing to do anyway instead of relying on the
alphabetical sorting of the image name.

For image /all this makes a nice list
For image /all-owner it is a bit messy but maybe more relevant

User can always pipe to sort and grep themselves.

Also:  Make it clear that Centos owner is the same as fedora owner

Signed-off-by: Bill Mills <wm.a.mills@gmail.com>