dockit: change run-interactive so it uses dockit_ecec2.sh
Directly executing at user level with the docker -u command does not set
the additional groups for the user.  This leaves out groups like sudo and
sudo_np.

As the first part of the fix for this make run-interactive use a second
trampoline script in the root.  Run this as root instead of try to run
as user.

Also explicitly save and return the rc of the docker exec process.

Signed-off-by: Bill Mills <wm.a.mills@gmail.com>