dockit: for sudo, fix interactive commands.
As the 2nd part of the fix for sudo, we use the dockit_exec2.sh trampoline
to run a command that switches to the user and then runs the specified
command.

This fixes sudo for shell and icmd dockit commands.

With these fixes we now verify that the return code from shell or icmd is
past back to the caller of the dockit command.  You can test using exit 0
or exit 1 (or 42) from the dockit shell command.

Signed-off-by: Bill Mills <wm.a.mills@gmail.com>