dockit: Support Ubuntu 24.04
Ubuntu 24.04 (and 23.X) has ubuntu user pre-installed as user 1000.
As we want to match the real user's name and UID this is a problem.
So if we detect the conflcit we delete the conflicting user.

Also in Ubuntu 24.04 (at least) the adduser package is not installed by
default so use the lower level usermod command to add the user to the
sudo_np group.  This is probably more distro portable anyway.

Signed-off-by: Bill Mills <wm.a.mills@gmail.com>