dockit: use existing container name if found
If the container for the distro & arch exists, use its name instead of
calculating one.

Due to symlinks and bind mounts, the PWD may have multiple apparent paths.
This can get things out of sync so let the first action to create the
container calculate the name and all others just use it.

Signed-off-by: Bill Mills <wm.a.mills@gmail.com>