dockit: disable the default setup function
The default setup function takes a while to run and may not be what people
want.  Disable it for now.

This is especially annoying right now as it is easy to create a new
container by mistake and it is hard to cancel one that is getting created
as ^C does not get passed as the setup it not interactive.

Both of those things should be fixed but this is the easiest to address
quickly.

Signed-off-by: Bill Mills <wm.a.mills@gmail.com>